refactor(sv-utils): keep pnpm namespace, hide detectPnpmMajor in pnpm-internals
Move detectPnpmMajor into a separate pnpm-internals.ts file that isn't
re-exported from index.ts. Restores the `export * as pnpm` namespace and
the `pnpm.allowBuilds` consumer call sites while keeping the version
detection helper out of the public API surface.
