+++ packages/sv-utils/src/index.ts
@@ -32,7 +32,7 @@ export * as json from './tooling/json.ts';
 export * as svelte from './tooling/svelte/index.ts';
 
 // Package manager helpers
-export { allowBuilds, onlyBuiltDependencies } from './pnpm.ts';
+export * as pnpm from './pnpm.ts';
 
 // Transforms — sv-utils = what to do to content, sv = where and when to do it.
 export { transforms } from './tooling/transforms.ts';
