+++ packages/sv/src/addons/tailwindcss.ts
@@ -1,4 +1,4 @@
-import { allowBuilds, transforms } from '@sveltejs/sv-utils';
+import { pnpm, transforms } from '@sveltejs/sv-utils';
 import { defineAddon, defineAddonOptions } from '../core/config.ts';
 
 const plugins = [
@@ -36,7 +36,7 @@ export default defineAddon({
 		sv.devDependency('tailwindcss', '^4.2.2');
 		sv.devDependency('@tailwindcss/vite', '^4.2.2');
 		if (packageManager === 'pnpm') {
-			sv.file(file.findUp('pnpm-workspace.yaml'), allowBuilds('@tailwindcss/oxide'));
+			sv.file(file.findUp('pnpm-workspace.yaml'), pnpm.allowBuilds('@tailwindcss/oxide'));
 		}
 
 		if (prettierInstalled) sv.devDependency('prettier-plugin-tailwindcss', '^0.7.2');
